The final stretch

+++ src/sodiumCompatibility/java/net/coderbot/iris/compat/sodium/impl/vertex_format/terrain_xhfp/XHFPTerrainVertex.java
@@ -60,7 +60,7 @@ public class XHFPTerrainVertex implements ChunkVertexEncoder, ContextAwareVertex
 
 	@Override
 	public long write(long ptr,
-					  Material material, Vertex vertex, int chunkId) {
+					  Material material, Vertex vertex, byte chunkId) {
 		uSum += vertex.u;
 		vSum += vertex.v;
 		vertexCount++;
@@ -69,7 +69,7 @@ public class XHFPTerrainVertex implements ChunkVertexEncoder, ContextAwareVertex
 		MemoryUtil.memPutShort(ptr + 2L, XHFPModelVertexType.encodePosition(vertex.y));
 		MemoryUtil.memPutShort(ptr + 4L, XHFPModelVertexType.encodePosition(vertex.z));
 		MemoryUtil.memPutByte(ptr + 6L, material.bits());
-		MemoryUtil.memPutByte(ptr + 7L, (byte) chunkId);
+		MemoryUtil.memPutByte(ptr + 7L, chunkId);
 
 		MemoryUtil.memPutInt(ptr + 8, vertex.color);
